+++ R/promises.R
@@ -27,7 +27,7 @@
 #' @details This function is an S3 method for the generic \code{as.promise} for
 #'     class 'mirai'.
 #'
-#'     Requires the packages \CRANpkg{promises} and \CRANpkg{later}.
+#'     Requires the \CRANpkg{promises} package.
 #'
 #'     Allows a 'mirai' to be used with the promise pipe \code{\%...>\%}, which
 #'     schedules a function to run upon resolution of the 'mirai'.